Add confirm for deleting comments.

+++ public/includes/js/commentsThingy.js
@@ -116,16 +116,17 @@ var deleteComment = function emptyDeleteComment() {
 	}
 
 	deleteComment = function deleteComment(id) {
-		API.DELETE("/comment/" + id, {},
-			function Success() {
-				var element = document.getElementById("comment_" + id);
-				element.parentNode.removeChild(element);
-
-			},
-			function Error() {
-				console.error("Failed to delete comment with ID " + id);
-			}
-		);
+		if (confirm("Are you sure you wish to delete this comment?")) {
+			API.DELETE("/comment/" + id, {},
+				function Success() {
+					var element = document.getElementById("comment_" + id);
+					element.parentNode.removeChild(element);
+				},
+				function Error() {
+					console.error("Failed to delete comment with ID " + id);
+				}
+			);
+		}
 	};
 
 	commentsThingy = function CommentsThingy(commentBodyId, root) {
